refactor(dashboard): extract groupByKey to shared module, remove dead param
- Extract groupByKey to common/collection.ts (was duplicated in store.ts
  and orchestra.ts)
- Remove unused _agentName parameter from buildAgentMotion
- Replace Set-based journal dedup with concat (no overlap possible)
